% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\title{Min Max Normalization}
\usage{
norm(data, na.rm = TRUE)
}
\arguments{
\item{data, }{which is the numeric data that you want to normalize, it could be a column, a vector}

\item{na.rm, }{specify whether you want to remove the NA data, the default option is true (move the data)}
}
\value{
a normalized data, the data type is the same with your input
}
\description{
Min Max Normalization
}
\details{
normalize the data using Min-Max Normalization scale
}
\examples{
vector_na <- c(1, 3, NA, 5)
vector_na_norm <- norm(vector_na)
vector_na_norm <- norm(vector_na, FALSE)
}
